+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# Unreleased
+- Add static code diagnostic double-literal-format
+
 # 1.4.0
 - Drop dependency on built_collection
 - Add set-exit-on-violation-level cli argument